+++ apps/web/app/(bookings)/bookings/single/page.tsx
@@ -1,7 +1,7 @@
-import Bookings from "@/module/Bookings"
 import { APP_NAME } from "@repo/constants"
 import { BOOKINGS } from "@/common/constants"
 import { Metadata } from "next"
+import SingleBooking from "@/module/Bookings/single"
 
 export const metadata: Metadata = {
   title: `${BOOKINGS} - ${APP_NAME}`,
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 }
 
 const BookingsPage = () => {
-  return <Bookings />
+  return <SingleBooking />
 }
 
 export default BookingsPage